pyproject.toml
[project]
name = "ogdc-runner"
authors = [
  { name = "Matt Fisher", email = "mfisher87@gmail.com" },
  { name = "Rushiraj Nenuji", email = "nenuji@nceas.ucsb.edu" },
  { name = "Trey Stafford", email = "trey.stafford@colorado.edu" },
  { name = "Robyn Marowitz", email = "robyn.marowitz@colorado.edu" },
]
description = "Defines OGDC recipe API(s) and submits recipes to OGDC for execution"
readme = "README.md"
license.file = "LICENSE"
requires-python = ">=3.11"
classifiers = [
  "Development Status :: 1 - Planning",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
  "click >=8",
  "hera >=5.18.0",
  "pydantic >=2.7.1",
  "pyyaml",
  "fsspec",
  "loguru",
]

[project.optional-dependencies]
# NOTE: "test" and "dev" are duplicated; why do we need both?
test = [
  "mypy ~=1.10.0",
  "pytest >=6",
  "pytest-cov >=3",
  "pytest-order",
  "types-pyyaml",
]
dev = [
  "mypy ~=1.10.0",
  "pytest >=6",
  "pytest-cov >=3",
  "types-pyyaml",
  "nox",
]
docs = [
  "sphinx>=7.0",
  "myst_parser>=0.13",
  "sphinx_copybutton",
  "sphinx_autodoc_typehints",
  "furo>=2023.08.17",
]

[project.urls]
Homepage = "https://github.com/qgreenland-net/ogdc-runner"
"Bug Tracker" = "https://github.com/qgreenland-net/ogdc-runner/issues"
Discussions = "https://github.com/qgreenland-net/ogdc-runner/discussions"
Changelog = "https://github.com/qgreenland-net/ogdc-runner/releases"


[project.scripts]
ogdc-runner = "ogdc_runner.__main__:cli"


[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"


[tool.hatch]
version.source = "vcs"
build.hooks.vcs.version-file = "src/ogdc_runner/_version.py"

[tool.hatch.envs.default]
features = ["test"]
scripts.test = "pytest {args}"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
  "error",
]
log_cli_level = "INFO"
testpaths = [
  "tests",
]
markers = [
  "order: mark test to run in a specific order"
]


[tool.coverage]
run.source = ["ogdc_runner"]
report.exclude_also = [
  '\.\.\.',
  'if typing.TYPE_CHECKING:',
]

[tool.mypy]
files = ["src", "tests"]
python_version = "3.11"
warn_unused_configs = true
strict = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
warn_unreachable = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "ogdc_runner.*"
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = [
  "jinja2.*",
  "kubernetes.*",
  "fsspec",
]
ignore_missing_imports = true


[tool.ruff]
src = ["src"]
target-version = "py311"

[tool.ruff.lint]
extend-select = [
  "B",        # flake8-bugbear
  "I",        # isort
  "ARG",      # flake8-unused-arguments
  "C4",       # flake8-comprehensions
  "EM",       # flake8-errmsg
  "ICN",      # flake8-import-conventions
  "G",        # flake8-logging-format
  "PGH",      # pygrep-hooks
  "PIE",      # flake8-pie
  "PL",       # pylint
  "PT",       # flake8-pytest-style
  "PTH",      # flake8-use-pathlib
  "RET",      # flake8-return
  "RUF",      # Ruff-specific
  "SIM",      # flake8-simplify
  "T20",      # flake8-print
  "UP",       # pyupgrade
  "YTT",      # flake8-2020
  "EXE",      # flake8-executable
  "NPY",      # NumPy specific rules
  "PD",       # pandas-vet
]
ignore = [
  "PLR09",    # Too many <...>
  "PLR2004",  # Magic value used in comparison
  "ISC001",   # Conflicts with formatter
  "PIE790",   # Unnecessary placeholder (`pass` and `...`)
  "T201",     # `print` found
  "EM102",    # Exception must not use an f-string literal, assign to variable first
  "RET504",   # Unnecessary assignment before return statement
]
isort.required-imports = ["from __future__ import annotations"]
# Uncomment if using a _compat.typing backport
# typing-modules = ["ogdc_runner._compat.typing"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
"noxfile.py" = ["T20"]
